Add read and write functionality to JSON file
This commit adds the ability to read and write data to a JSON file using the serde
and std::io modules. A `Person` struct is defined to represent the data, and a vector
of `Person` structs is created to populate the file.

The `data.json` file is created using `File::create()`, and then written to using
`serde_json::to_writer_pretty()`. To read the data back in, the file is opened using
`File::open()`, and then read into a vector of `Person` structs using
`serde_json::from_reader()`.

This commit is a complete implementation of the program, and includes proper error
handling and formatting.

Signed-off-by: [Sabry Awad] <dr.sabry1997@gmail.com>